requirements.txt
google-adk
uvicorn
langchain
langchain-community
langchain-google-vertexai
langchain-google-genai
faiss-cpu
python-dotenv
google-cloud-storage
unstructured
google-cloud-aiplatform 
requests>=2.31.0
